+++ steerclear/views.py
@@ -1,15 +1,2 @@
-from steerclear import app
-from flask import render_template
-
-"""
-heartbeat
----------
-Simple check to see if server is running
-"""
-@app.route('/')
-def heartbeat():
-        return "pulse"
-
-@app.route('/index')
-def index():
-    return render_template('index.html')
+from steerclear.api.views import *
+from steerclear.driver_portal.views import *
